+++ rl_toolkit/networks/actor_critic.py
@@ -1,0 +1,132 @@
+from rl_toolkit.networks.layers import Actor, Critic
+from tensorflow.keras import Model
+
+import tensorflow as tf
+
+
+class ActorCritic(Model):
+    """Combines the actor and critic into an end-to-end model for training."""
+
+    def __init__(
+        self,
+        num_of_outputs: int,
+        **kwargs
+    ):
+        super(ActorCritic, self).__init__(**kwargs)
+
+        # Actor
+        self.actor = Actor(num_of_outputs)
+
+        # Critic 1
+        self.critic_1 = Critic()
+        self.critic_1_target = Critic()
+        self._train_target(self.critic_1, self.critic_1_target, tau=tf.constant(1.0))
+
+        # Critic 2
+        self.critic_2 = Critic()
+        self.critic_2_target = Critic()
+        self._train_target(self.critic_2, self.critic_2_target, tau=tf.constant(1.0))
+
+        # init param 'alpha' - Lagrangian constraint
+        self.log_alpha = tf.Variable(0.0, trainable=True, name="log_alpha")
+        self.alpha = tf.Variable(0.0, trainable=False, name="alpha")
+        self.target_entropy = tf.cast(
+            -num_of_outputs, dtype=tf.float32
+        )
+
+    def train_step(self, data):
+        # Re-new noise matrix every update of 'log_std' params
+        self.actor.reset_noise()
+
+        next_action, next_log_pi = self.actor(data["next_observation"])
+        _, log_pi = self.actor(data["observation"])
+
+        # update 'Alpha'
+        self.alpha.assign(tf.exp(self.log_alpha))
+        with tf.GradientTape() as tape:
+            losses = -1.0 * (
+                self.log_alpha * tf.stop_gradient(log_pi + self.target_entropy)
+            )
+            alpha_loss = tf.nn.compute_average_loss(losses)
+
+        gradients = tape.gradient(alpha_loss, [self.log_alpha])
+        self.optimizer.apply_gradients(zip(gradients, [self.log_alpha]))
+
+        # target Q-values
+        next_Q_values = tf.minimum(
+            self.critic_1_target(
+                [data["next_observation"], next_action]
+            ), 
+            self.critic_2_target(
+                [data["next_observation"], next_action]
+            )
+        )
+
+        # Bellman Equation
+        Q_targets = tf.stop_gradient(
+            data["reward"]
+            + (1.0 - data["terminal"])
+            * self.gamma
+            * (next_Q_values - self.alpha * next_log_pi)
+        )
+
+        # update 'Critic 1'
+        with tf.GradientTape() as tape:
+            Q_values = self.critic_1(
+                [data["observation"], data["action"]]
+            )
+            losses = tf.losses.huber(  # less sensitive to outliers in batch
+                y_true=Q_targets, y_pred=Q_values
+            )
+            Q1_loss = tf.nn.compute_average_loss(losses)
+
+        gradients = tape.gradient(Q1_loss, self.critic_1.trainable_variables)
+        self.optimizer.apply_gradients(
+            zip(gradients, self.critic_1.trainable_variables)
+        )
+
+        # update 'Critic 2'
+        with tf.GradientTape() as tape:
+            Q_values = self.critic_2(
+                [data["observation"], data["action"]]
+            )
+            losses = tf.losses.huber(  # less sensitive to outliers in batch
+                y_true=Q_targets, y_pred=Q_values
+            )
+            Q2_loss = tf.nn.compute_average_loss(losses)
+
+        gradients = tape.gradient(Q2_loss, self.critic_2.trainable_variables)
+        self.optimizer.apply_gradients(
+            zip(gradients, self.critic_2.trainable_variables)
+        )
+
+        # update 'Actor'
+        with tf.GradientTape() as tape:
+            # predict action
+            y_pred, log_pi = self.actor(data["observation"])
+
+            # predict q values
+            Q_values = tf.minimum(self.critic_1([data["observation"], y_pred]), self.critic_2([data["observation"], y_pred]))
+
+            losses = self.alpha * log_pi - Q_values
+            actor_loss = tf.nn.compute_average_loss(losses)
+
+        grads = tape.gradient(actor_loss, self.actor.trainable_variables)
+        self.actor.optimizer.apply_gradients(
+            zip(grads, self.actor.trainable_variables)
+        )
+
+        # Soft update target networks
+        self._train_target(self.critic_1, self.critic_1_target, tau=self.tau)
+        self._train_target(self.critic_2, self.critic_2_target, tau=self.tau)
+
+        return {"critic_loss": (Q1_loss + Q2_loss), "actor_loss": actor_loss, "alpha_loss": alpha_loss}
+
+    def _train_target(self, source, target, tau):
+        for source_weight, target_weight in zip(
+            source.trainable_variables, target.trainable_variables
+        ):
+            target_weight.assign(tau * source_weight + (1.0 - tau) * target_weight)
+
+    def call(self, inputs):
+        pass