+++ rl_toolkit/networks/layers/__init__.py
@@ -1,0 +1,3 @@
+from .actor import Actor  # noqa
+from .critic import Critic  # noqa
+from .noise import MultivariateGaussianNoise  # noqa